[FIX] FIX issue with default pricelist on partner
Two issue are fix here, an random issue in the test and an real issue
for the default pricelist on the partner.

When the stock module is installed before the module account the
configuration of the default pricelist is not the same with the demo
data. Indeed stock module add a company and so when we install the
chart of account genereric as we are in multi-company, and as we change
the currency of the main company a property configuration is added for
 each company see code here:
https://github.com/odoo/odoo/blob/10.0/addons/product/models/res_company.py#L41
This is why we need to force in the demo data to set the right
pricelist in order to be sure to always have the same.

Also when we create a new customer from the site we need to use the
right pricelist configured on the shopinvader backend