[IMP] selector should propose similare variant
The selector have a concept of constraint.
The order of this constraint depend on the sequence of the attribute
For example when you select a color the size is constrained.
But the size can not constraint the color (front end recurcivity that
block the user selection)
But changing the color should avoid changing the selected size.
This commit introduction "preference" concept. So when we change the
color we go (if available) to the page with the same size.

This improve usability