[IMP] mechanism to set session return "out-of-band"
The locomotive frontend interprets
store_cache and set_session values in result.  Since those are not really
data returned by the service but
some sort of out-of band data,
it is not always practical to ask
each service (which may not know
about such information) to set
it just before their "return" statement.
With this mechanism it is possible
to set this session data anywhere,
and it is injected in the response
automatically.