s_customer_multi_user: refactor is_invader_user
Before this change the check was done
based on a new partner type 'invader_client_user'.

This was a limitation because:

1. you could not use other partner types for the same record

2. invader users created via backend and not having the special type
   (like most of the cases when you have an existing set of partners)
   were not acknowledged as invader user and they couldn't see
   all the addresses from the main account

Now the flag is computed at binding level
and the default type for the partner is 'contact'.

No migration step required since the selection value will be
automatically deleted from db when upgrading the module,
and the new value will be computed the 1st time is required.