s_sale_packaging: do not drop uom qty
Dropping the uom qty is not needed
because it will get recomputed anyway
when passing a pkg qty.
BUT if you pop it and you pass pkg qty = 0
to drop the packaging value
the qty won't be the one that you might expect
from the request sent w/ the field 'item_qty'.