[IMP] shopinvader_sale_coupon : cast bool skip_coupon_recompute
problem from our client
the cart update_item methods requires skip_coupon_recompute to be a boolean while the rest (item_id, item_qty) can be string. We are facing troubles due to this constraint, it will be better to accept string for the skip_coupon_recompute as well. So we won’t need to cast the value every where we use the method